pyproject.toml
[project]
name = "freelance-tesla"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "pandas-stubs>=2.3.3.251219",
    "streamlit>=1.52.2",
]
